EWayBill Details integration

+++ OmkarFCUBEAPI/AdminMasters/Models/EWayBillModel.cs
@@ -55,11 +55,11 @@
         public int pincode_of_consignee { get; set; }
         public string state_of_supply { get; set; }
         public string actual_to_state_name { get; set; }
-        public int total_invoice_value { get; set; }
+        public decimal total_invoice_value { get; set; }
         public int taxable_amount { get; set; }
         public int cgst_amount { get; set; }
         public int sgst_amount { get; set; }
-        public int igst_amount { get; set; }
+        public decimal igst_amount { get; set; }
         public int cess_amount { get; set; }
         public string transporter_id { get; set; }
         public string transporter_name { get; set; }
@@ -100,6 +100,4 @@
         public string transporter_document_date { get; set; }
         public string group_number { get; set; }
     }
-
-
 }
